Unified the types for 'content'
standardized the props for the components
finished the component wrapper
