+++ frontend/src/components/TextArea/TextArea.tsx
@@ -1,16 +1,9 @@
 import React from 'react';
 import './TextArea.css';
-import { QData, BindFieldFunction, CollectAnswerFunction } from '../../types/common';
+import { CommonComponentAttributes } from '../../types/CommonComponentAttributes';
 
 type TextAreaProps = {
-    attributes: {
-        q: QData;
-        bindField: BindFieldFunction;
-        collectAnswer: CollectAnswerFunction;
-        content: {
-            errorMessage: string;
-        };
-    };
+    attributes: CommonComponentAttributes;
 };
 export function TextArea(props: TextAreaProps) {
     const {
