+++ frontend/src/components/TextInput/TextInput.tsx
@@ -1,15 +1,8 @@
 import React from 'react';
 import './TextInput.css';
-import { QData, BindFieldFunction, CollectAnswerFunction } from '../../types/common';
+import { CommonComponentAttributes } from '../../types/CommonComponentAttributes';
 type TextInputProps = {
-    attributes: {
-        q: QData;
-        bindField: BindFieldFunction;
-        collectAnswer: CollectAnswerFunction;
-        content: {
-            errorMessage: string;
-        };
-    };
+    attributes: CommonComponentAttributes;
 };
 export function TextInput(props: TextInputProps) {
     const {
