+++ frontend/src/compositions/Footer/Footer.tsx
@@ -2,12 +2,11 @@ import React from 'react';
 import ciitLogo from '../../data/images/CiiT Logo.png';
 import ceticpLogo from '../../data/images/CET-ICP Logo.png';
 import './Footer.css';
+import { ContentText } from '../../types/ContentText';
 
+type FooterContent = Pick<ContentText, 'footerText1' | 'footerText2'>;
 type FooterProps = {
-    content?: {
-        footerText1?: string;
-        footerText2?: string;
-    };
+    content: FooterContent;
 };
 export function Footer(props: FooterProps) {
     const { content = { footerText1: '', footerText2: '' } } = props;
