--- frontend/src/compositions/WorkshopScreening/Modal/types.ts
@@ -1,9 +1,0 @@
-export type ModalContent = {
-    screeningDateMarried: string;
-    screeningDate: string;
-    modalText1: string;
-    modalText2: string;
-    modalText3: string;
-    modalText4: string;
-    modalExitButton: string;
-};
