+++ frontend/src/compositions/WorkshopScreening/QuestionTypes.ts
@@ -1,3 +1,4 @@
+import { ContentText } from "../../types/ContentText";
 import { BindFieldFunction, CollectAnswerFunction } from "../../types/common";
 
 export type QuestionProps_Q = {
@@ -12,9 +13,6 @@ export type QuestionProps = {
     q?: QuestionProps_Q | undefined | null; 
     bindField: BindFieldFunction; 
     setQuestion: (value: string | null) => void;  
-    content: {
-        errorMessage: string;
-        screeningProceedButton: string;
-    }; 
+    content: ContentText; 
     collectAnswer: CollectAnswerFunction;
 };
