+++ frontend/src/types/CommonComponentAttributes.ts
@@ -1,0 +1,10 @@
+import { ContentText } from "./ContentText";
+import { Question } from "./Question";
+import { BindFieldFunction, CollectAnswerFunction } from "./common";
+
+export type CommonComponentAttributes<TContent = ContentText> = {
+    bindField: BindFieldFunction;
+    collectAnswer: CollectAnswerFunction;
+    q: Question;
+    content: TContent;
+};
