+++ frontend/src/types/Question.ts
@@ -1,0 +1,11 @@
+import { FormElementName } from "../utilities/formElements";
+
+export type Question = {
+    category: string;
+    required?: boolean;
+    parentQuestionSlug?: string;
+    answerSelections?: string;
+    slug: string;
+    text: string;
+    questionType: FormElementName;
+};
